prints the sizes of various datatype on the computer

+++ 0x00-hello_world/6-size.c
@@ -6,15 +6,15 @@
 */
 int main(void)
 {
-	int i;
+	long int i;
 	char c;
 	float f;
 	long long int j;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(j));
+	printf("Size of a char: %lu byte(s)\n", (unsigned)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", (unsigned)sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned)sizeof(i));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned)sizeof(j));
 	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
